Flask-MySQL/requirements.txt
flask
sqlalchemy
pymysql
docusign-esign
pyjwt
requests
python-dotenv